Создание модулей и точки входа
